Extract optionalAuth into separate middleware file
