+++ server/src/middleware/auth-guard.ts
@@ -22,16 +22,3 @@ export const authGuard = (allowedRoles?: UserRole[]) => async (req: Request, res
 
   return next();
 };
-
-export const optionalAuth = async (req: Request, res: Response, next: NextFunction) => {
-  const session = await auth.api.getSession({
-    headers: fromNodeHeaders(req.headers),
-  });
-
-  if (session) {
-    req.session = session.session;
-    req.user = session.user;
-  }
-
-  return next();
-};
