+++ server/src/routes/job-posting-router.ts
@@ -2,8 +2,9 @@ import express from 'express';
 import { JobPostingController } from '../controllers/job-posting-controller.ts';
 import { container } from 'tsyringe';
 import { TOKENS } from '../config/dependency-tokens.ts';
-import { authGuard, optionalAuth } from '../middleware/auth-guard.ts';
+import { authGuard } from '../middleware/auth-guard.ts';
 import { USER_ROLE } from '../data/util/constants.ts';
+import { optionalAuth } from '../middleware/optional-auth.ts';
 
 export const getJobPostingRouter = () => {
   const router = express.Router();
@@ -19,11 +20,7 @@ export const getJobPostingRouter = () => {
 
   router.post('/', authGuard([USER_ROLE.RECRUITER]), jobPostingController.createJobPosting);
 
-  router.patch(
-    '/:id',
-    authGuard([USER_ROLE.RECRUITER, USER_ROLE.ADMIN]),
-    jobPostingController.updateJobPosting,
-  );
+  router.patch('/:id', authGuard([USER_ROLE.RECRUITER, USER_ROLE.ADMIN]), jobPostingController.updateJobPosting);
 
   router.delete('/:id', jobPostingController.deleteJobPosting);
